unit test for books api

+++ src/main/java/com/projects/book_api/services/BookService.java
@@ -37,7 +37,11 @@ public class BookService {
     }
 
     public void deleteBook(Long id) {
-        bookRepository.deleteById(id);
+        if (bookRepository.existsById(id)) {
+            bookRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Book not found");
+        }
     }
 
 }
